requirements.txt
hikari[speedups]
aiosqlite
hikari-arc
uvloop
python-dotenv